requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon
pandas
apache-airflow-providers-snowflake
apache-airflow-providers-slack==8.1.0
